chore: rename remaining Anki files to make each one's role clear
Continuing the rename from d7113bb, same strategy for the last two
ambiguous file names:

- src/multiFieldApkg.js -> src/politicianNoteTypeSql.js (defines the
  Deutschland:Politiker note type and the raw SQL collection/note/card
  writer that works around anki-apkg-export's two-field limit)
- src/ankiExport.res -> src/apkgWriterFacade.res (thin ReScript
  bindings over politicianNoteTypeSql.js plus anki-apkg-export's own
  addMedia/save/writeFileSync, mirroring the existing cheerioFacade.res
  naming pattern for a bindings module)

Updated the AnkiExport.* call sites in ankiDeckBuilder.res to
ApkgWriterFacade.*, the spec file's import, and the README section
that names both files. No behavior change; res:build regenerates
cleanly and 75/75 tests pass.
