+++ README.md
@@ -39,6 +39,6 @@ Produces `output/<file>.apkg` next to the input JSON, importable into Anki/AnkiD
 Every deck shares one note type, `Deutschland:Politiker` (fields: Name, Partei,
 Amt/Ministerium, Profil-Photo; two card directions — role → person and photo → person).
 The fields and card templates are defined once, as the `POLITICIAN_FIELDS` and
-`POLITICIAN_TEMPLATES` constants in [`src/multiFieldApkg.js`](./src/multiFieldApkg.js) —
-that's the single place to edit if the layout needs to change; `src/ankiExport.res` only
+`POLITICIAN_TEMPLATES` constants in [`src/politicianNoteTypeSql.js`](./src/politicianNoteTypeSql.js) —
+that's the single place to edit if the layout needs to change; `src/apkgWriterFacade.res` only
 binds into it, it doesn't build any template itself.
